+++ design-system/tag/examples/constellation/tag-colors.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Tag from '../../src/tag/simple-tag';
+import { SimpleTag as Tag } from '../../src';
 
 export default () => (
   <>