+++ design-system/tag/examples/constellation/tag-max-length.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Tag from '../../src/tag/simple-tag';
+import { SimpleTag as Tag } from '../../src';
 
 const cupcakeipsum = 'Croissant tiramisu gummi bears.';