+++ design-system/tag/examples/constellation/tag-removable-link.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Tag from '../../src/tag/removable-tag';
+import Tag from '../../src';
 
 export default () => (
   <Tag