+++ design-system/tag/examples/constellation/tag-removable.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
 
-import Tag from '../../src/tag/removable-tag';
+import Tag from '../../src';
 
 export default () => <Tag text="Removable tag" removeButtonLabel="Remove" />;