+++ design-system/tag/src/tag/internal/removable/ert.ts
@@ -1,0 +1,3 @@
+import { RemovableTagProps } from './index';
+
+export default function ertHack(_: RemovableTagProps) {}