+++ design-system/tag/src/tag/internal/shared/ert.ts
@@ -1,0 +1,3 @@
+import { SimpleTagProps } from './types';
+
+export default function ertHack(_: SimpleTagProps) {}