+++ growth/feature-flag-client/src/tracked-flag.ts
@@ -1,5 +1,11 @@
 import { FlagShape, Flag, CustomAttributes } from './types';
-import { isBoolean, isObject, isOneOf, isString } from './lib';
+import {
+  isBoolean,
+  isObject,
+  isOneOf,
+  isString,
+  validateFlagExplanation,
+} from './lib';
 
 export default class TrackedFlag implements Flag {
   flagKey: string;
@@ -10,6 +16,11 @@ export default class TrackedFlag implements Flag {
     flag: FlagShape,
     exposureData?: CustomAttributes,
   ) => void;
+  sendAutomaticExposure: (
+    flagKey: string,
+    value: string | boolean | object,
+    flagExplanation?: FlagShape['explanation'],
+  ) => void;
 
   constructor(
     flagKey: string,
@@ -19,11 +30,17 @@ export default class TrackedFlag implements Flag {
       flag: FlagShape,
       exposureData?: CustomAttributes,
     ) => void,
+    sendAutomaticExposure: (
+      flagKey: string,
+      value: string | boolean | object,
+      flagExplanation?: FlagShape['explanation'],
+    ) => void,
   ) {
     this.flagKey = flagKey;
     this.value = flag.value;
     this.trackExposure = trackExposure;
     this.flag = flag;
+    this.sendAutomaticExposure = sendAutomaticExposure;
   }
 
   getBooleanValue(options: {
@@ -31,15 +48,21 @@ export default class TrackedFlag implements Flag {
     shouldTrackExposureEvent?: boolean;
     exposureData?: CustomAttributes;
   }): boolean {
-    if (!isBoolean(this.value)) {
-      return options.default;
+    let value = options.default;
+
+    if (isBoolean(this.value)) {
+      value = this.value as boolean;
     }
+
+    const flagExplanation = validateFlagExplanation(this.flag, options.default);
 
     if (options.shouldTrackExposureEvent) {
       this.trackExposure(this.flagKey, this.flag, options.exposureData);
     }
 
-    return this.value as boolean;
+    this.sendAutomaticExposure(this.flagKey, value, flagExplanation);
+
+    return value;
   }
 
   getVariantValue(options: {
@@ -48,24 +71,32 @@ export default class TrackedFlag implements Flag {
     shouldTrackExposureEvent?: boolean;
     exposureData?: CustomAttributes;
   }): string {
-    if (
-      !isString(this.value) ||
-      !isOneOf(this.value as string, options.oneOf)
-    ) {
-      return options.default;
+    let value = options.default;
+
+    if (isString(this.value) && isOneOf(this.value as string, options.oneOf)) {
+      value = this.value as string;
     }
+
+    const flagExplanation = validateFlagExplanation(this.flag, options.default);
+
     if (options.shouldTrackExposureEvent) {
       this.trackExposure(this.flagKey, this.flag, options.exposureData);
     }
 
-    return this.value as string;
+    this.sendAutomaticExposure(this.flagKey, value, flagExplanation);
+
+    return value;
   }
 
   getJSONValue(): object {
-    if (!isObject(this.value)) {
-      return {};
+    let value = {};
+
+    if (isObject(this.value)) {
+      value = this.value as object;
     }
 
-    return this.value as object;
+    this.sendAutomaticExposure(this.flagKey, value);
+
+    return value;
   }
 }